+++ packages/next/src/data/store.ts
@@ -6,7 +6,7 @@ import type { SpreeNextOptions } from '../types';
  * Get the current store configuration.
  */
 export async function getStore(options?: SpreeNextOptions): Promise<StoreStore> {
-  return getClient().store.get({
+  return getClient().store.store.get({
     locale: options?.locale,
     currency: options?.currency,
   });